Added authorization checks for editing and deleting posts per user

+++ controllers/feed.js
@@ -49,7 +49,6 @@ exports.createPost = (req, res, next) => {
     const title = req.body.title;
     const content = req.body.content;
     let creator;
-    //TODO: create post in DB
     const post = new Post({
         title: title,
         content: content,
@@ -127,6 +126,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -158,7 +162,11 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
-            //TODO: Check logged in user and post author
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized.');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId)
                 .then(result => {
